+++ api/endpoints/models/search_models.py
@@ -1,11 +1,17 @@
 from pydantic import BaseModel
-from typing import List
-
+from typing import List, Optional
+from .general_models import Limits
 
 """
 Search results are not yet working so the below needs to 
 be updated accordingly
 """
+
+
+class SearchInput(BaseModel):
+    search_string: str
+    language: str = ""
+    limits: Optional[Limits]
 
 
 class SearchResults(BaseModel):
