+++ api/endpoints/models/utils_models.py
@@ -1,5 +1,13 @@
 from pydantic import BaseModel
-from typing import List
+from typing import List, Optional
+from .general_models import Limits
+
+
+class CountMatchesInput(BaseModel):
+    file_name: str = ""
+    score = 0
+    par_length = 0
+    limits: Optional[Limits]
 
 
 class CountMatchesOutput(BaseModel):
